+++ src/components/TopBar/TopBar.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styles from './TopBar.module.scss'
 
-import API, { staticData } from '../../utils/api'
+import API, { staticData, defineCurrentTerm } from '../../utils/api'
 import { Comparables } from '../../domain/Comparable'
 import Degree from '../../domain/Degree'
 import Course from '../../domain/Course'
@@ -32,6 +32,7 @@ import Brightness5Icon from '@material-ui/icons/Brightness5'
 import i18next from 'i18next'
 import Menu from '@material-ui/core/Menu'
 import Avatar from '@material-ui/core/Avatar'
+import Typography from '@material-ui/core/Typography'
 
 class TopBar extends React.Component <{
 	showAlert: (message: string, severity: 'success' | 'warning' | 'info' | 'error' | undefined) => void
@@ -68,6 +69,10 @@ class TopBar extends React.Component <{
 	}
 
 	async componentDidMount(): Promise<void> {
+		// Update term
+		const currTermId = await defineCurrentTerm()
+		this.onSelectedAcademicTerm(currTermId)
+
 		const degrees = await API.getDegrees()
 		this.setState({
 			degrees: degrees ?? []
@@ -301,9 +306,10 @@ class TopBar extends React.Component <{
 								<Icon>help</Icon>
 							</IconButton>
 						</Tooltip>
-						{/* <IconButton color='inherit' onClick={() => {this.setState({settingsDialog: true})}} component="span">
+						<IconButton color='inherit' onClick={() => {this.setState({settingsDialog: true})}} component="span"
+							style={{display: 'none'}}>
 							<Icon>settings</Icon>
-						</IconButton> */}
+						</IconButton>
 					</Toolbar>
 				</AppBar>
 				<Dialog open={this.state.settingsDialog}
@@ -331,6 +337,9 @@ class TopBar extends React.Component <{
 								)}
 							</Select>
 						</FormControl>
+						<Typography variant="caption" gutterBottom style={{marginTop: '8px'}}>
+							Isto irá apagar todos as disciplinas e turnos selecionados.
+						</Typography>
 					</DialogContent>
 					<DialogActions>
 						<div />