+++ CHANGELOG.md
@@ -6,11 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.0] - 02/09/2021
+
 ### Added
+
+- Settings dialog to change semester (hidden for now)
 
 ### Fixed
 
 - Remove courses that can't be obtained
+- Degrees had not been fetched with the updated academic term
 
 ### Removed